refactor(app-management): 移除无用的 app-list 组件，将应用编辑的抽屉和表单组件移动到对应的页面目录下
