+++ packages/react-app/src/pages/admin/app-management/components/app-edit-drawer.tsx
@@ -8,7 +8,7 @@ import { useEffect, useState } from 'react'
 import { AppDetailDrawerModeEnum } from '@/enums'
 import { useAuth } from '@/hooks/use-auth'
 
-import SettingForm from './setting-form'
+import SettingForm from './app-setting-form'
 
 interface IAppEditDrawerProps extends DrawerProps {
 	detailDrawerMode: AppDetailDrawerModeEnum
